Add Contact constructor
Add constructor to Contact class. Along with the constructor, a helped function called calculateAge calculates a person's age by parsing their birthDate into a Date object and comparing it to the current date (also as a Date object). There is also some test code in Main to ensure that the constructor works properly.